src/pyproject.toml
[project]
name = "uk-retirement-planner"
version = "0.1.0"
description = "UK Retirement Planner API"
requires-python = ">=3.13"
dependencies = [
    "fastapi<0.110.0",
    "pydantic<2.0.0",
    "PyJWT>=2.8.0",
    "cryptography>=42.0.0"
]

[dependency-groups]
dev = []

[tool.pywrangler]
main = "main.py"
